+++ src-tauri/src/core/service.rs
@@ -456,12 +456,12 @@ impl ServiceManager {
         Self(ServiceStatus::Unavailable("Need Checks".into()))
     }
 
-    pub const fn config() -> Option<clash_verge_service_ipc::IpcConfig> {
-        Some(clash_verge_service_ipc::IpcConfig {
+    pub const fn config() -> clash_verge_service_ipc::IpcConfig {
+        clash_verge_service_ipc::IpcConfig {
             default_timeout: Duration::from_millis(30),
             retry_delay: Duration::from_millis(250),
             max_retries: 6,
-        })
+        }
     }
 
     pub async fn init(&mut self) -> Result<()> {