+++ src-tauri/src/utils/resolve/mod.rs
@@ -153,7 +153,7 @@ pub(super) async fn init_verge_config() {
 }
 
 pub(super) async fn init_service_manager() {
-    clash_verge_service_ipc::set_config(ServiceManager::config()).await;
+    clash_verge_service_ipc::set_config(Some(ServiceManager::config())).await;
     if !is_service_ipc_path_exists() {
         return;
     }