+++ Changelog.md
@@ -3,6 +3,7 @@
 ### 🐞 修复问题
 
 - Linux 无法切换 TUN 堆栈
+- macOS service 启动项显示名称(试验性修改)
 
 <details>
 <summary><strong> ✨ 新增功能 </strong></summary>
@@ -10,6 +11,7 @@
 - **Mihomo(Meta) 内核升级至 v1.19.16**
 - 支持连接页面各个项目的排序
 - 实现可选的自动备份
+- 连接页面支持查看已关闭的连接（最近最多 500 个已关闭连接）
 
 </details>